removes UniformLayout
Uniform Layout was a one size fits all approach to Uniforms. This was
removed in favor of breaking out DescriptorSetLayouts, DescriptorPools,
and specific Uniform Data (such as Dynamic Uniform Buffers and samplers)
into vectors of each data type. The specific Uniform Data is now in a
std::variant of DescriptorType collections, so that each collection can
handle allocating/updating/deleting/etc uniforms in it's own way.